+++ dart/lib/src/transport/http_transport.dart
@@ -7,9 +7,7 @@ import 'noop_encode.dart' if (dart.library.io) 'encode.dart';
 import '../noop_client.dart';
 import '../protocol.dart';
 import '../sentry_options.dart';
-import '../utils.dart';
 import '../sentry_envelope.dart';
-
 import 'transport.dart';
 import 'rate_limiter.dart';
 
@@ -35,7 +33,10 @@ class HttpTransport implements Transport {
 
   HttpTransport._(this._options, this._rateLimiter)
       : _dsn = Dsn.parse(_options.dsn!),
-        _headers = _buildHeaders(_options.sdk.identifier) {
+        _headers = _buildHeaders(
+          _options.platformChecker.isWeb,
+          _options.sdk.identifier,
+        ) {
     _credentialBuilder = _CredentialBuilder(
       _dsn,
       _options.sdk.identifier,
@@ -163,7 +164,7 @@ class _CredentialBuilder {
   }
 }
 
-Map<String, String> _buildHeaders(String sdkIdentifier) {
+Map<String, String> _buildHeaders(bool isWeb, String sdkIdentifier) {
   final headers = {'Content-Type': 'application/x-sentry-envelope'};
   // NOTE(lejard_h) overriding user agent on VM and Flutter not sure why
   // for web it use browser user agent