+++ flutter/test/sentry_flutter_options_test.dart
@@ -1,13 +1,13 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:sentry/src/platform/platform.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/sentry_flutter_options.dart';
 
 void main() {
   group('SentryFlutterOptions', () {
-    testWidgets('auto breadcrumb tracking', (WidgetTester tester) async {
-      final options = SentryFlutterOptions();
-      options.configureBreadcrumbTrackingForPlatform(TargetPlatform.android);
+    testWidgets('auto breadcrumb tracking: has native integration',
+        (WidgetTester tester) async {
+      final options = SentryFlutterOptions(checker: MockPlatformChecker(true));
 
       expect(options.enableAppLifecycleBreadcrumbs, isFalse);
       expect(options.enableWindowMetricBreadcrumbs, isFalse);
@@ -15,34 +15,18 @@ void main() {
       expect(options.enableTextScaleChangeBreadcrumbs, isFalse);
       expect(options.enableMemoryPressureBreadcrumbs, isFalse);
       expect(options.enableAutoNativeBreadcrumbs, isTrue);
+    });
 
-      options.configureBreadcrumbTrackingForPlatform(TargetPlatform.iOS);
+    testWidgets('auto breadcrumb tracking: without native integration',
+        (WidgetTester tester) async {
+      final options = SentryFlutterOptions(checker: MockPlatformChecker(false));
 
-      expect(options.enableAppLifecycleBreadcrumbs, isFalse);
-      expect(options.enableWindowMetricBreadcrumbs, isFalse);
-      expect(options.enableBrightnessChangeBreadcrumbs, isFalse);
-      expect(options.enableTextScaleChangeBreadcrumbs, isFalse);
-      expect(options.enableMemoryPressureBreadcrumbs, isFalse);
-      expect(options.enableAutoNativeBreadcrumbs, isTrue);
-
-      // for all other platform the inverse is true
-      final platforms = [
-        TargetPlatform.fuchsia,
-        TargetPlatform.linux,
-        TargetPlatform.macOS,
-        TargetPlatform.windows,
-      ];
-
-      for (final platform in platforms) {
-        options.configureBreadcrumbTrackingForPlatform(platform);
-
-        expect(options.enableAppLifecycleBreadcrumbs, isTrue);
-        expect(options.enableWindowMetricBreadcrumbs, isTrue);
-        expect(options.enableBrightnessChangeBreadcrumbs, isTrue);
-        expect(options.enableTextScaleChangeBreadcrumbs, isTrue);
-        expect(options.enableMemoryPressureBreadcrumbs, isTrue);
-        expect(options.enableAutoNativeBreadcrumbs, isFalse);
-      }
+      expect(options.enableAppLifecycleBreadcrumbs, isTrue);
+      expect(options.enableWindowMetricBreadcrumbs, isTrue);
+      expect(options.enableBrightnessChangeBreadcrumbs, isTrue);
+      expect(options.enableTextScaleChangeBreadcrumbs, isTrue);
+      expect(options.enableMemoryPressureBreadcrumbs, isTrue);
+      expect(options.enableAutoNativeBreadcrumbs, isFalse);
     });
 
     testWidgets('useFlutterBreadcrumbTracking', (WidgetTester tester) async {
@@ -69,4 +53,28 @@ void main() {
       expect(options.enableAutoNativeBreadcrumbs, isFalse);
     });
   });
+}
+
+class MockPlatformChecker implements PlatformChecker {
+  MockPlatformChecker(this.hasNativeIntegrationValue);
+
+  final bool hasNativeIntegrationValue;
+
+  @override
+  bool get hasNativeIntegration => hasNativeIntegrationValue;
+
+  @override
+  bool isDebugMode() => true;
+
+  @override
+  bool isProfileMode() => false;
+
+  @override
+  bool isReleaseMode() => false;
+
+  @override
+  bool get isWeb => false;
+
+  @override
+  Platform get platform => throw UnimplementedError();
 }