+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Unreleased
 
 * Fix: `Sentry.close()` closes native SDK integrations (#388)
+* Feat: Support for macOS (#389)
+* Feat: Support for Linux (#402)
+* Feat: Support for Windows (#407)
 * Fix: Mark `Sentry.currentHub` as deprecated (#406)
+* Fix: Set console logger as default logger in debug mode (#413)
 * Fix: Use name from pubspec.yaml for release if package id is not available (#411)
 * Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests (#414)
+* Bump: sentry-cocoa to v7.0.0 (#424)
+* Feat: Support for Out-of-Memory-Tracking on macOS/iOS (#424)
+* Fix: Trim `\u0000` from Windows package info (#420)
+* Fix: `dist` was read from `SENTRY_DSN`, now it's read from `SENTRY_DIST` (#442)
 
 ## Breaking Changes: